# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: >-
    Suivi des conventions de partenariat de la Région Souss-Massa.
    Le statut d'alerte n'est jamais stocké : il est recalculé à chaque requête
    par rapport à la date du jour.
servers:
  - url: /api
    description: Base API path
security:
  - apiSession: []
...
paths:
...
  /admin/users:
    get:
      operationId: listUsers
      tags: [admin]
      summary: Lister les comptes (administrateur uniquement)
      responses:
        "200":
          description: Comptes sans informations de mot de passe
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/AdminUser" }
    post:
      operationId: createUser
      tags: [admin]
      summary: Créer un compte (administrateur uniquement)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UserInput" }
      responses:
        "201":
          description: Compte créé
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminUser" }
        "400": { description: Données invalides }
  /reference-data:
    get:
      operationId: getReferenceData
      tags: [reference-data]
      summary: Référentiel courant de l'organigramme et des maîtrises d'ouvrage
      responses:
        "200":
          description: Référentiel hiérarchique courant
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ReferenceData" }
        "401": { description: Authentification requise }
  /admin/reference-data/{kind}:
    post:
      operationId: createReferenceDataEntry
      tags: [admin]
      summary: Ajouter une entrée au référentiel (administrateur uniquement)
      parameters:
        - name: kind
          in: path
          required: true
          schema: { $ref: "#/components/schemas/ReferenceDataKind" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ReferenceDataInput" }
      responses:
        "200":
          description: Référentiel complet mis à jour
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ReferenceData" }
        "400": { description: Données ou parent invalides }
        "403": { description: Administrateur uniquement }
        "409": { description: Une entrée de même nom existe déjà }
  /admin/reference-data/{kind}/{id}:
    patch:
      operationId: renameReferenceDataEntry
      tags: [admin]
      summary: Renommer une entrée sans la supprimer (administrateur uniquement)
      parameters:
        - name: kind
          in: path
          required: true
          schema: { $ref: "#/components/schemas/ReferenceDataKind" }
        - name: id
          in: path
          required: true
          schema: { type: integer, minimum: 1 }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ReferenceDataRename" }
      responses:
        "200":
          description: Référentiel complet mis à jour
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ReferenceData" }
        "400": { description: Nom invalide ou conflit d'unicité }
        "403": { description: Administrateur uniquement }
        "404": { description: Entrée introuvable }
        "409": { description: Une entrée de même nom existe déjà }
  /admin/users/{id}:
    patch:
      operationId: updateUser
      tags: [admin]
      summary: Modifier un compte (administrateur uniquement)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UserUpdate" }
      responses:
        "200":
          description: Compte modifié
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminUser" }
        "400": { description: Données invalides }
        "404": { description: Compte introuvable }
    delete:
      operationId: deleteUser
      tags: [admin]
      summary: Supprimer définitivement un compte non administrateur
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, minimum: 1 }
      responses:
        "204": { description: Compte supprimé et sessions révoquées }
        "400": { description: Identifiant invalide }
        "403": { description: Les comptes administrateurs ne peuvent pas être supprimés }
        "404": { description: Compte introuvable }
  /admin/users/{id}/password:
    post:
      operationId: resetUserPassword
      tags: [admin]
      summary: Réinitialiser le mot de passe et révoquer les sessions
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PasswordInput" }
      responses:
        "204": { description: Mot de passe réinitialisé }
        "404": { description: Compte introuvable }
  /admin/audit:
    get:
      operationId: listAudit
      tags: [admin]
      summary: Consulter l'historique des modifications (administrateur uniquement)
      parameters:
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
        - name: offset
          in: query
          schema: { type: integer, minimum: 0, default: 0 }
        - name: conventionId
          in: query
          schema: { type: integer, minimum: 1 }
      responses:
        "200":
          description: Historique paginé
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuditPage" }
...
components:
  securitySchemes:
    apiSession:
      type: apiKey
      in: cookie
      name: region_session
  schemas:
...
    ReferenceDataKind:
      type: string
      enum: [directions, divisions, services, services-dgs, maitrises-ouvrage]
    ReferenceService:
      type: object
      required: [id, nom]
      properties:
        id: { type: integer }
        nom: { type: string }
    ReferenceDivision:
      type: object
      required: [id, nom, services]
      properties:
        id: { type: integer }
        nom: { type: string }
        services:
          type: array
          items: { $ref: "#/components/schemas/ReferenceService" }
    ReferenceDirection:
      type: object
      required: [id, direction, divisions]
      properties:
        id: { type: integer }
        direction: { type: string }
        divisions:
          type: array
          items: { $ref: "#/components/schemas/ReferenceDivision" }
    ReferenceData:
      type: object
      required: [organisation, servicesDgs, maitrisesOuvrage]
      properties:
        organisation:
          type: array
          items: { $ref: "#/components/schemas/ReferenceDirection" }
        servicesDgs:
          type: array
          description: Services directement rattachés à la Direction Générale des Services, sans direction ni division intermédiaire.
          items: { $ref: "#/components/schemas/ReferenceService" }
        maitrisesOuvrage:
          type: array
          items: { $ref: "#/components/schemas/ReferenceService" }
    ReferenceDataInput:
      type: object
      required: [nom]
      properties:
        nom: { type: string, minLength: 1, maxLength: 255 }
        parentId:
          type: integer
          minimum: 1
          description: ID de direction pour une division ou de division pour un service; interdit pour les directions, services-dgs et maîtrises d'ouvrage.
    ReferenceDataRename:
      type: object
      required: [nom]
      properties:
        nom: { type: string, minLength: 1, maxLength: 255 }
    Role:
      type: string
      enum: [admin, directeur, directeur_general_services, chef_division, chef_service]
    AuthUser:
      type: object
      required: [id, username, role, direction, division, service, active]
      properties:
        id: { type: integer }
        username: { type: string }
        role: { $ref: "#/components/schemas/Role" }
        direction: { type: ["string", "null"] }
        division: { type: ["string", "null"] }
        service: { type: ["string", "null"] }
        active: { type: boolean }
    AdminUser:
      allOf:
        - $ref: "#/components/schemas/AuthUser"
        - type: object
          required: [createdAt]
          properties:
            createdAt: { type: string, format: date-time }
...
    UserInput:
      type: object
      required: [username, password, role]
      properties:
        username: { type: string, minLength: 1, maxLength: 100 }
        password: { type: string, minLength: 8, maxLength: 1024 }
        role: { $ref: "#/components/schemas/Role" }
        direction: { type: ["string", "null"] }
        division: { type: ["string", "null"] }
        service: { type: ["string", "null"] }
        active: { type: boolean, default: true }
    UserUpdate:
      type: object
      properties:
        username: { type: string, minLength: 1, maxLength: 100 }
        role: { $ref: "#/components/schemas/Role" }
        direction: { type: ["string", "null"] }
        division: { type: ["string", "null"] }
        service: { type: ["string", "null"] }
        active: { type: boolean }
    PasswordInput:
      type: object
      required: [password]
      properties:
        password: { type: string, minLength: 8, maxLength: 1024 }
    AuditEvent:
      type: object
      required: [id, conventionId, action, actorUsername, before, after, createdAt]
      properties:
        id: { type: integer }
        conventionId: { type: integer }
        action: { type: string, enum: [create, update, delete] }
        actorUsername: { type: string }
        before: { type: ["object", "null"], additionalProperties: true }
        after: { type: ["object", "null"], additionalProperties: true }
        createdAt: { type: string, format: date-time }
    AuditPage:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/AuditEvent" }
        total: { type: integer }