# lib/api-zod/src/generated/types/referenceData.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Suivi des conventions de partenariat de la Région Souss-Massa. Le statut d'alerte n'est jamais stocké : il est recalculé à chaque requête par rapport à la date du jour.
 * OpenAPI spec version: 0.1.0
 */
import type { ReferenceDirection } from './referenceDirection';
import type { ReferenceService } from './referenceService';

export interface ReferenceData {
  organisation: ReferenceDirection[];
  /** Services directement rattachés à la Direction Générale des Services, sans direction ni division intermédiaire. */
  servicesDgs: ReferenceService[];
  maitrisesOuvrage: ReferenceService[];
}