# lib/api-zod/src/generated/types/referenceDataInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Suivi des conventions de partenariat de la Région Souss-Massa. Le statut d'alerte n'est jamais stocké : il est recalculé à chaque requête par rapport à la date du jour.
 * OpenAPI spec version: 0.1.0
 */

export interface ReferenceDataInput {
  /**
     * @minLength 1
     * @maxLength 255
     */
  nom: string;
  /**
     * ID de direction pour une division ou de division pour un service; interdit pour les directions, services-dgs et maîtrises d'ouvrage.
     * @minimum 1
     */
  parentId?: number;
}