# lib/api-zod/src/generated/types/referenceDataKind.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Suivi des conventions de partenariat de la Région Souss-Massa. Le statut d'alerte n'est jamais stocké : il est recalculé à chaque requête par rapport à la date du jour.
 * OpenAPI spec version: 0.1.0
 */

export type ReferenceDataKind = typeof ReferenceDataKind[keyof typeof ReferenceDataKind];


export const ReferenceDataKind = {
  directions: 'directions',
  divisions: 'divisions',
  services: 'services',
  'services-dgs': 'services-dgs',
  'maitrises-ouvrage': 'maitrises-ouvrage',
} as const;